Add: midpoint and linear function
Introduced 2 functions:
  cudf::interpolate::midpoint
  cudf::interpolate::linear

Add test case foe midpoint
liner is not tested well
