--- cpp/src/quantiles/quantiles.h
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) 2018, NVIDIA CORPORATION.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#pragma once
-
-//Quantile (percentile) functionality
-
-#include "cudf.h"
-
-struct QuantiledIndex {
-  size_t lower_bound;
-  size_t upper_bound;
-  size_t nearest;
-  double fraction;
-};
-
-QuantiledIndex find_quantile_index(gdf_size_type length, double quant);
-
-
